+++ routes/characters.js
@@ -3,8 +3,8 @@ const router = express.Router({ mergeParams: true });
 const { verifyLogin, isOwner, countCharacters } = require('../utilities/middleware');
 const characters = require('../controllers/characters')
 
-router.post('/', verifyLogin, isOwner, countCharacters, characters.addCharacter)
+router.post('/:characterId', verifyLogin, characters.addCharacter)
 
-router.delete('/:characterId', verifyLogin, isOwner, characters.removeCharacter)
+router.delete('/:characterId', verifyLogin, characters.removeCharacter)
 
 module.exports = router;