+++ CHANGELOG.md
@@ -2,10 +2,14 @@
 All notable changes to this project will be documented in this file. 
 
 ## Unreleased
+### Added
+- Character to team from character search results.
+- `getCharacter` utility function
 ### Changed
 - Updated `README.md`
 - Pushed team list further below map.
-
+### Removed
+- Add character form from show team page
 
 ## feature-31 // 2021-04-26
 ### Added